refactor: Component 디렉토리 분리
