+++ src/frontend/src/components/notice/NoticeType.vue
