+++ src/frontend/src/components/question/QuestionDetail.vue
